Refactor ethereum-lido-v3 raw slot indexing
