--- protocols/substreams/ethereum-lido-v3/src/modules/mod.rs
@@ -1,8 +1,0 @@
-#[path = "1_map_protocol_components.rs"]
-mod map_protocol_components;
-
-#[path = "2_store_protocol_state.rs"]
-mod store_protocol_state;
-
-#[path = "3_map_protocol_changes.rs"]
-mod map_protocol_changes;
